add helper functions for contour detection
and a small application to tune the thresholds of the algorithm

The application looks prints the detected contours in the image. What contours
are detected can be tuned by modifying the thresholds with the sliders.

A bounding box id drawn around the largest contour with 4 sides.

The polar coordinates from the center or the image to the center of the largest tetrahedron
is printed on the image. Note that down is the positive y direction because of y coordinates in image pixels.