increase rate at which pictures are taken
by moving the take picture call to the beginning of the callback instead of the end

Also, changed to use a custom callback triggered off the picture taken notification, so that
there are fewer times the callback is called when there isn't a picture ready for download.

The period between pictures seems to be about 1.2s now.