Add basic "flight mode" to center over a target
It takes the trajectory from the center of the image to the target
and command the pitch and roll accordingly

The pitch and roll angles are scaled according to how far the target is from the
center of the image

this is about the most naive controls approach possible. observationally,
I'm seeing the vehicle way overshoot its mark so I probably have to command
it for less than 1s at a time.