use convex hull algorithm to get a better idea of the target
use the open cv to get the contours, but just getting the largest contour
wasn't alway reliable. If the whole landing square was not in the image, to
contour wouldn't detect four points.

Took the points from the approximate contour and use the convex hull algorithm.
This allowed be to detect the landing pad even if the entire pad was not in the
frame.

Also, use the average brightness of the area inside the convex hull to determine if
the target is in fact a white sheet of paper.

Create a Target data class to contain more info about the target in the frame, and
piped that though.

Also, added code to command the drone to hold still while tracking if it isn't actively trying to
make a move towards the target.