improve the target detection algorithm
The algorithm no longer just considers the largest convex hull.

The algorithm looks at the average brightness of all the convext hulls,
and then going from brightest to least bright, looks to see if there is
a right angle in the hull.

The reasoning here is that the white paper target should be the brightest,
and there should almost always be a corner visible on the target. The case where
this might not be true is if an edge of the target is only visible in the corner of the image.
I'm going to ignore this case for now.

Also, improve the graphing of convex hulls.

Add a flag to tune edge detection to optionally turn on graphing the convex hulls